fixed student priv

+++ docs/install/db/atutor_schema.sql
@@ -114,7 +114,7 @@ CREATE TABLE `course_enrollment` (
   `member_id` mediumint(8) unsigned NOT NULL default '0',
   `course_id` mediumint(8) unsigned NOT NULL default '0',
   `approved` enum('y','n','a') NOT NULL default 'n',
-  `privileges` smallint(5) unsigned NOT NULL default '0',
+  `privileges` INT(10) unsigned NOT NULL default '0',
   `role` varchar(35) NOT NULL default '',
   `last_cid` mediumint(8) unsigned NOT NULL default '0',
   PRIMARY KEY  (`member_id`,`course_id`)
